split google photos auth and media handling
